2026/text-to-mp4/requirements.txt
flask
moviepy
pillow
openai
